[expect/gtm8223] Fix for when test ran in UTF8 mode
The recent comment (https://gitlab.com/YottaDB/DB/YDBTest/-/merge_requests/1418) fixed some issues with this
test but moved the expect output into the reference file. Unfortunately, mupip complains when ^%GI is fed via
heredoc from an expect script, the output file has an error about a mismatch in modes between the input and
the run mode. This error was not a problem when we were only grepping for PASS in the expect output file
but now that the sanitized output file is part of the reference file, we just alter the reference file to
exclude that warning if NOT running in UTF8.