[YDBTest#420/GTM-9238] Add test v63012/gtm9238 to test $ZSTRPLLIM functionality - issue with it not causing STPOFLOW when it should
Not all of the release note for this issue was understood. For clarity, here's the original GT.M release note:
```
GT.M manages detection of STPCRIT errors to minimize the chance that repeats of those errors
indicate decreasing amounts of available space; previously repeated STPCRIT errors
inappropriately indicated a loss of string pool space. Also, an attempt to set the limit lower
than the miminum string pool size adds the requested limit to the minimum size; previously,
GT.M silently ignored such attempts. (GTM-9238)
```
This test does check that we don't get a neverending set of STPCRIT errors without a fatal STPOFLOW
occuring but we never really understood what the "errors indicate decreasing amounts of available
space" part meants. In addition, we decided that the "add the minimum to the requested amount"
part caused unexpected results so made it always set the value minimum if the user attempted
to set it lower. This new setting of the mininum is verified in this test.