[expect/gtm8223] Fix issue with output lines being split due to short terminal line size
This test was modified to add the expectinit.exp initialization to the expect script in
this issue. This initialization set the line length parameter appropriately but the test
did ANOTHER terminal spawn after that for the test to run in so all the defaults returned
and ignored the line length change. Removing that extra spawn solved the issue.