[YDBTest#441][v63013/gtm8793] Test GTM-8793 that syslog gets an EXITSTATUS error if the procstuckexec script fails
This test intentionally creates a MUTEXLCKALERT situation which causes the procstuckexec (if defined) to be driven.
This situation is created by running DSE in a pipeline and doing a 'crit -seize' to grab the critical section. Then
we start another pipe-controlled process that runs YDB doing an update to a var and thus hanging waiting for crit.
Wait for 30 seconds for MUTEXLCKALERT timeout, then start checking the syslog every 5 seconds for a max of 70 seconds
(90 seconds total) looking for our messages. Our primary focus is the EXITSTATUS message but we also look for the
other two messages that are seen with it (STUCKACT and MUTEXLCKALERT) and put all three of these messages into the
reference file.

Also, we have defined the procstuckexec envvar (ydb_procstuckexec) to a non-existent script which causes an error
to be returned which is what causes the EXITSTATUS error to show up.