[sudo/gtm7759] Changes to allow test to run on SUSE
The test ran fine under Ubuntu and Redhat but needed changes to run under SUSE. Three issues appeared:

1. If the ydbtest3 group or userid existed (perhaps left over from a failed run or other unexpected event),
the test would fail when it tried to create them. The test was made tolerant of this condition.
2. The 'userdel' command outputs the line "no crontab entry for ydbtest3" when it is working normally. To
solve, I redirected output from the userdel command to a file like we do for useradd.
3. The USERGROUPS_ENAB flag defined in /etc/login.defs is enabled in both Ubuntu and Redhat but is
disabled in SUSE so the ydbtest3 group was not being removed after the test ran. Added a section to see
if the group still exists and removes it if still there to solve this issue.