[#463] Fixed sudo test failure due to ydb1187 subtest addition
Issue:
-
`sudo` test failed with following diff during in-house testing:
```
7d6

< PASS from gtm7759

8a8

> PASS from gtm7759

--------- Subtest level diff follows ---------
```

While fixing the above issue, also found a problem in `sudo/instream.csh`. Separate lists were not being created for replication and non-replication subtests.

Reason:
-
The sudo test failed due to recent changes made to the test system in !1437. The new subtest (ydb1187) was added to the wrong list essentially changing the order of the subtests causing problems with the output reference file.

Fix:
-
Moved the subtest `ydb1187` from the common list to the non replic list of subtests to be run under sudo test.

Additionally, `sudo/instream.csh` has been updated to set and run the correct subtests depending on whether the `-replic` option is supplied or not, as initially it was running all of the subtests regardless.