[YDBTest#460 v63013/gtm9311] Test to validate GTM-9311
The ^%YGBLSTAT routine no longer overwrites the 'x' and 'd' variable trees when called
from another routine. Previously the x and d variables were overwritten and/or had subscripted
entries added to it. The change is the x and d variables are now NEWed so are saved and
restored over the call to ^%YGBLSTAT.