[#461 v63013/gtm9313] Add test to verify $ORDER() functions properly
The #461 issue was about $ORDER() returning a potentially invalid value when:
1. The variable is subscripted.
2. At least one subscript contains a boolean expression.
3. The expression must also reference a global variable.
4. The 2nd argument to $ORDER() should be a literal.

In this case, it can happen that an invalid value was returned. When this test
is run under a version prior to V63012, it usually fails in one of the early
attempts if not the first. This fails on r1.34 which is based on V63011.

Note, this test reworks the boolean expression generator from v60000/gtm7277 and
its boolgen.m routine. This routine was broken into pieces that can be called
individually. There's no use for it elsewhere at this time but if it happens that
this is useful for other tests, it will be renamed and moved to the com directory
instead.