[#439/#440 GTM-8772, GTM-8784, GTM-9308] Adds test v63013/gtm8772ANDgtm8784 testing MUPIP cmds while DB is FROZEN
This test has 6 basic phases:
1. Run the complete list of MUPIP SET commands while the DB is FROZEN -ONLINE while a process has the DB open (no standalone access).
2. Run the complete list of MUPIP SET commands while the DB is FROZEN -ONLINE while DB is NOT open (has standalone access).
3. Run the complete list of MUPIP SET commands while the DB is UNFROZEN while DB is NOT open (has standalone access).
4. Run a MUPIP RUNDOWN command while the DB is FROZEN -ONLINE
5. Run a MUPIP BACKUP command while the DB is FROZEN -ONLINE. Also specified the noauto-release option to prevent backup from releasing the freeeze.
6. Run five MUPIP SET commands specifically mentioned in the issue that they only flush the DB file-header - not any buffers that may be pending. In this case, there is no database freeze.

The testing of GTM-9308 occurs in phase 2 where standalone access is available. When MUPIP tries get standalone access, it notices that the shared memory already exists and assumes this to be a damaged database. This is incorrect as the shared memory was left because of the freeze. So GTM-9308 (which is part of V63014) is being pulled forward so it is merged as part of V63013. Without that fix, phase 2 above gets incorrect errors (MUUSERECOV). All of the standalone-required commands fail with this error without GTM-9308.