[#442 GTM-8838] Add v63013/gtm8838 test to verify WFR/BUS/BTS statistics are being reported
Making the new statistics (WFR, BUS, and BTS) have non-zero values is difficult as these stats
cover rare conditions so this test does not try to do that. Instead, it verifies these the 3
statistics mentioned in the title are being provided as well as verifying that one cannot
access these variables with their old names and instead are available via ^%PEEKBYNAME and/or
MUPIP DUMPFHEAD.

Additionally, this test uses a new white-box case to set all of the counters in gvstats_rec_t
to be the offset of that field in the gvstats_rec_t structure. These values are then validated
by pulling the stats via both ^%PEEKBYNAME() and MUPIP DUMPFHEAD.