[YDBTest#445 GTM-9230] Add v63013/gtm9230 test to verify non-TP NOISOLATION works
We discovered that YottaDB and GTM don't do as good a job with NOISOLATION and non-TP restarts as they do
with TP restarts. So just going off the number of restarts between the two modes was not satisfactory
as a test. What we decided to do is run a test that would benefit from NOISOLATION with
gtm_nontprestart_log_first and gtm_nontprestart_log_delta set to 1 to cause non-TP restart messages to be
sent to the syslog. We then extract those messages and validate the restart type. In an ISOLATION environment,
the type of the restarts we get should be t_blkmod_t_end3 (13) but in a NOISOLATION environment, the restart
type should be t_blkmod_t_end4 (14). Both of these are verified.