[YDBTest#449] Add v63013/gtm9277 test to validate fixes to gtm_side_effect mode intrinsic calls
Release note pertaining to this issue:
>The GT.M compiler, when using gtm_side_effects, deals appropriately cases when the last argument in a Boolean expression has a side effect and the result of the entire expression is evaluated as an integer. Most integer evaluations are for intrinsic function arguments. The side effects include extrinsic expressions, $INCREMENT() and indirect evaluations. Subsequent to the introduction of the gtm_side_effects mode, such evaluations could fail with an ASSERTPRO fatal error, segmentation violation (SIG-11), or rarely an incorrect result. (GTM-9277)

This issue being tested here is when crafted boolean expressions are used for integer arguments
in intrinsic calls. The problem occurs when the argument in question is being fed to the underlying
C call as an integer instead of as an mval. The crafted boolean expression needs to have as its
last term a term that has a side-effect (generally an extrinsic call, an indirect or a $increment()).

This test has a list of expressions that test these types of arguments and generates routines for
each expression with each routine having statements that test the expression in 3 ways (currently):
1. RHS of a SET command.
2. Argument of an IF command.
3. Used as a command prefix.

The test then drives each of the generated routines noting failures.