[YDBTest#446/GTM-9252] Add v63013/gtm9252 test for read-only DBs not closing correctly at rundown
This test uses ^%PEEKBYNAME() on a fileheader field from two synchronized processes. When both of
them rundown, the last process out sends a SYCALL error for semctl(IPC_RMID,nnn). This means the DB
does not close correctly. While this error was fixed a long time ago in YottaDB (YDB#150) the issue
is still a problem in V63013 though is fixed in V63014. We created the test now, have verified it
fails on V63012/V63013 and works on V63014 and it works in YottaDB.