[YDBTest#466] Modify v63013/gtm7277 to use com/BooleanExprGen.m
The issue YDBTest#466 also needed a boolean generator so the boolean generator for gtm7277
was taken out, generalized and made available to any test in the com directory. This issue
(YDBTest#466) was created to convert gtm7277 to use this common version instead of the
embedded version it was using thus fixes to one place fix potential issues in multiple tests.

This test also makes a minor change to com/BooleanExprGen.m to both document its entry points
and add the ability to define additional extrinsic routines (which is what gtm7277 uses it for)
or to add other code or entry points to the end of the generated routine.

Also moved the setting of an 'exprcnt' variable out of Init^BooleanExprGen and into gtm7277.m
where it belongs.