[YDB#924] New sudo/ydb924 test for testing termination of ydbinstall/ydbinstall.sh if not run as root unless --dry-run is specified
This commit add a new test `sudo/ydb924` which tests that the `ydbinstall/ydbinstall.sh` script terminates if it is not run by root unless the --dry-run option is specified.

The following changes have been tested:

 - YDB#924 changes:
   - Running `ydbinstall/ydbinstall.sh` as root with --dry-run
   - Running `ydbinstall/ydbinstall.sh` as root
   - Running `ydbinstall/ydbinstall.sh` as non root with --dry-run
   - Running `ydbinstall/ydbinstall.sh` as non root

A test for YDB#765 has not been added to this commit since there is no good way to currently test it.