[#438] Fix v63013/gtm7987 parsing of receiver server log messages
This test starting failing November first because the 'day digits' in the receiver server
log file timestamp went from `Oct 31` to `Nov  1` - note the extra space for alignment in
the latter. The verify routine was parsing the log using $zpiece() with a space as a delimiter
so this additional space caused it to extract the wrong fields so it never found what it was
looking for.

This change modifies the parsing algorithm to use `' : '` as the delimiter so we bypass the
entire timestamp and get a consistent record value to look at.