[#468] Add v63014/gtm7628 test to verify source and receiver servers can use 64GB jnlpools
The catch is, the system has to have enough memory for this to run so this test adds to the
test system framework the JNLPOOL64GB_DISALLOW flag so the test can be bypassed when the system
does not have enough memory (tested in the test system by new ydb_allow64GB_jnlpool envvar).

This update also modified `com/process.awk` to get rid of a lot of the wildcard references that
initially caused problems with the original keyword we added 'NOALLOW64GB_JNLPOOL'. This matched
the /NO.?JNL/ regex causing intermittent errors in runs. While that is no longer an issue with
the new name (JNLPOOL64GB_DISALLOW) we use now, we decided to fix the abundant use of wildcards
anyway and keep the list to a mostly fixed set of names.

Note - this test provides test coverage for YDB#946 as it depends on the journal pool size being
reported correctly as a 64 bit value which it was not prior to YDB#946.