[YDB#779] [YDBTest#483] Add test v63014/gtm9329 to test a couple of $ZTIMEOUT fixes in V6.3-014
The two fixes (and subissues in this test) are the following:
1. Disabling the current (expired) timeout from within the timeout code works properly - previously to loop (dbg only).
2. Trying to assign an invalid code vector should not replace the current working vector - previously left blank

The first subissue is tested with a white box test that is only available in V6.3-014 releases. It simulates the same
error the issue complains about and shows that this issue no longer occurs. Since white box tests are only available
on debug builds, this stage of the subtest is restricted to debug builds. This subissue fails on V63013 but only because
the lines put out by the white box test do not appear. Since the window that the signal has to pop in is so small, is
is beyond difficult to make this test actually fail in the manner described by the release note.

The second subissue is tested in a second stage. This subissue fails on V63013 and earlier releases with the code vector
not being properly maintained across the failure (the previous valid vector should have been maintained). Unlike the
first subissue, this subissue fails the same with V63013 or earlier the same regardless of pro or debug so this stage
of the subtest is not restricted to debug builds.