[YDB#935][#490] New r136/ydb935 tests for Python multi-threaded applications
We have two separate issues in YottaDB that are due to Python's
multi-threaded code. Python, due to its use of the Global Interpreter
Lock (GIL) ensures that we only have a single process accessing YottaDB
at a single moment. Thus it uses the Simple Unthreaded API. However, we
found a couple of issues due to the use this API:

1. [Fixed in YDB!1220] Ensure that timer resets happen on the correct thread
2. [Fixed in YDB!1274] Remove incorrect assert when there is a SimpleAPI application with multiple threads

(1) is tested by `r136/inref/ydb935-1.py` which is the by and large
unmodified test case from @RamSailopal (who disclaimed copyright here:
https://gitlab.com/YottaDB/DB/YDBTest/-/issues/490#note_1193527342).

(2) is tested by `r136/inref/ydb935-2.py`.

ASAN/CLANG test fails with error `Your application is linked against
incompatible ASan runtimes` for unknown reasons we would rather not
spend time on, so we disable the test if we find that we are running on
a YottaDB build with ASAN/Clang