[#483] Fix v63014/gtm9329 test case to ignore time portion of $ZTIMEOUT when comparing $ZTIMEOUT values
Part of this test was to verify that the $ZTIMEOUT value did not change when attempting to set an
a $ZTIMEOUT value with a bad code vector. This comparison was made directly of the entire $ZTIMEOUT
value. Did not realize at the time of writing that the time portion of the $ZTIMEOUT value actually
counted down the time until the $ZTIMEOUT was driven so unless the before and after check of the
$ZTIMEOUT value around the attempt to set an invalid value occurred within 1ms, the value would
not match.

The fix was to compare only the code vector (not including the timer) when checking if $ZTIMEOUT was
modified when attempting to set an invalid code vector. Also, do some minor range checks on the
timer value (should be between 5 and 0).