[YDB#779] [#473] Add r136/gtm8863a test to verify 3 fileheader fields migrated properly with db_auto_upgrade()
When merging GTM-8863 in V6.3-013 into the YottaDB master branch for the upcoming r1.36 release,
some YottaDB added fileheader fields had to move from where they were located in r1.34 because
upstream used the space for something else. This test is to verify those 3 fields were migrated
correctly by db_auto_upgrade() when the file is first opened by the new version.

The 3 fields are as follows:
- max_procs.time
- max_procs.cnt
- reorg_sleep_nsec

The test basically creates an r1.34 database, does a few commands to make these 3 values somewhat unique
the saves the output of a DSE dump -fileheader. Then switches to the current test release, and once again
does a DSE dump -fileheader then checks the values of the three fields in question and verifies them.