[YDB#779] [#474] Add v63014/gtm8863b test of new toggle type statistics
The merge of V6.3-014 into YottaDB added the following new toggle statistics:
  (DEXA, GLB, JNL, MLK, PRC, TRX, ZAD, JOPA, AFRA, BREA, MLBA & TRGA)

These statistics are either 0 or 1 in a given process with the '1' values denoting that the
sitation described by the statistic name (these names are described in the ZSHOW help with
the other statistics) is true. Global views will show the sum of these toggle values across
the process subset involved.

The new test runs a set of 12 processes that do database busy work with some use of M locks
and makes a few of these (mainly MLK, TRX, MLBA, and TRGA) turn non-zero. The main process
runs ^%YGBLSTAT() every 3 seconds for 15 seconds. The main process takes each line from
^%YGBLSTAT() and looks at each value generated. We need two things for the test to pass:

1. Each of the "monitored stats" (in this case MLK and MLBA as TRX and TRGA are not consistent
   enough) must have a non-zero value in one of the 6 fetches of toggle stat values.
2. Each value should not exceed the maximum number of worker processes (12 by default).

With these conditions, the test passes on both our x86 and ARM based systems.