[YDB#779] [YDB#986] [#469] Add new test v63014/gtm8800 to verify changes to MUPIP FTOK/SEMAPHORE
This test has several parts/pieces:

1. Verify that all shared segment ids reported by MUPIP FTOK are correct (both DBs and jnlpool/recvpool).
2. Verify that all semaphore ids reported by MUPIP FTOK are correct (both DBs and jnlpool/recvpool).
3. Verify that the -ID parameter works.
4. Verify that the -noheader parameter works.
5. Verify that the -only parameter works.
6. Verify can provide a list of repl instance files with -jnlpool and/or -recvpool
7. Verify MUPIP SEMAPHORE can deal with a list of semaphores (used to verify existence of all
semaphores gathered for item #2).

We accomplish 1 and 2 by executing commands to retrieve all the database, jnlpool, and recvpool
semids and shmids and then running a module in each instance that verifies the appropriate ids
are in use by each of those instances. Also, as noted, MUPIP SEMAPHORE validates the existence
of all those semaphores.

The remainder of the items are executed but with the particulars of their output largely ignored
since it is being validated in other ways. The only piece of this issue that is not validated is
the actual ftok itself because, as YDBTest#469 says, YottaDB has its own internal ftok algorithm
that has little in common with the system ftok so we have no way to externally derive it.

Note that this test uses updates to the $$^%PEEKBYNAME() facility made in YDB#986 (YDB!1313) so
has a dependency on them. This usage is also a test for those additions.