[YDB#935][#490] Fix r136/ydb935 tests for RHEL 7
The trick used to wait for output in a file to show up (tail -f/sed q)
does not work on RHEL 7 (This is hinted at here:
https://unix.stackexchange.com/questions/304987/why-does-sed-q-work-differently-when-reading-from-a-pipe;
look for CentOS 7).

I changed this to use (tail -f ) | yottadb -r %XCMD 'for  read/quit'.
Tested on RHEL 7, RHEL 8, and SLED 15 with 10 runs each.