[YDB#779][#489] New v63014/gtm9338 test to check that the $gtm_tmp environment variable can control the placement of files used in trigger compilation
This commit adds a new test `v63014/gtm9338` which checks that trigger compilation responds to the location specified by the gtm_tmp environment variable.

The test assigns a value to $gtm_tmp then creates a database. The first trigger definition file is created and added (using MUPIP TRIGGER). This trigger is successfully added. Then the write permissions for the $gtm_tmp directory are revoked. Another trigger definition file is created and added. This trigger addition is unsuccessful. This shows that trigger compilation responds to the location specified by $gtm_tmp instead of using the default value of /tmp.