Fix r136/ydb935 subtest failure when YottaDB is built with GCC/ASAN
We have observed curl stall on Ubuntu 22.04 systems in the ydb935 test
when YottaDB is built with GCC and ASAN.

We confirmed empirically that curl hangs due to the use of LD_PRELOAD,
which was applied globally to every process in the test script (invoking
python3 in the test requires this env var set in order to correctly
invoke the YDBPython wrapper). The solution is to only apply LD_PRELOAD
to enable ASAN on the Python process, and no others.