[YDB#979] New r138/ydb979 subtest to test MUPIP FTOK of .repl and .dat file names 255 bytes or more long or .repl file(s) not in current working directory
The bulk of YDB#979 is to fix an issue with multiple replication instance files specified with -jnlpool
and/or -recvpool and all files showing up having the same FTOK and FileId. Since we cannot really validate
the FTOK or FileId values, this new test has a different way of testing them. It creates and moves into a
subdirectory so the current working directory does not have any of the replication files being queried and
then verifies that the FTOK returned is not -1 meaning it found the file.

While YDB#979 was being reviewed, a situation was noted and a test was requested for the situation where
the value of $gtm_repl_instance (whether specified to the MUPIP FTOK command or defaulted via the envvar),
was exactly 255 bytes long to make sure that command did not get into issues (see
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1308#note_1241464986 for need for a test and see
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1308#note_1244775945 for discussion on how the earlier
code failed).

Similarly, when a database name was specified that was 255 bytes or longer, it could overflow the buffer
and cause a segfault error. Note that both the case of the filename being exactly 255 bytes long or if the
filename was longer that 255 could both cause overflow issues but for different reasons. In the longer case,
the buffer was being overflowed but in the equal-to-255-bytes case, it caused the strncpy() call that was
loading the buffer to not put a null terminator on the end leading to running past the buffer end when it
was next utilized.