[YDB#779] [YDBTest#485] Add new v63014/gtm9332 to test the rollover of pctl->ctl->wakeups with white box test
When the pctl->ctl->wakeups wakeup counter (promoted in V63014 from 32 to 64 bit unsigned value)
rolled over (to 0) with V63013 and earlier releases, it could lead to the process returning and $TEST set
so the user thought they had acquired the lock when THEY DID NOT.

The WBTEST_LCKWAKEOVRFLO white box test simulates this condition by always setting pctl->ctl->wakeups
to one of two values. The first iteration, when pctl->ctl->wakeups starts off at 0, when the lock is
released, the value will be set to 0xFFFFFFFF. The second time, it will be set to 0xFFFFFFFFFFFFFFFF
which is where an increment will roll it to 0 which is where the problem area occurs.

When in this condition, the issue says we could experience a positive lock indication without actually
getting the lock so we verify that each of 3 tries to get the lock actually does acquire the lock. We
do this in a couple of different processes causing processes to be awakened from waiting for the lock
which is where the issue occurs. This test has each process grab the lock, then drive an LKE command
in a pipe where we can fetch which PID holds the lock (if any does). We verify that the current process
that just did the lock command has the lock so either no lock or wrong process with the lock are both
detected errors.

This test fails with V63014 with GTM-9322 reverted and the assert(retval) near the top
of mlk_lock() allows the assert to pass even if retval is 0 if the WBTEST_LCKWAKEOVRFLO white-box test
is activated so we can get into the errant lock situation described above.