Fix com/online_reorg.csh to exclude GTM error messages too (not just corresponding YDB messages)
The com/online_reorg.csh script excludes the errors MUTRUNCNOV4 and MUTRUNCNOTBG from the output
of MUPIP REORG when copying the output logs into the reference file. This was done in commit
be4dd38d which changed the GTM messages to be excluded to YDB messages. Since we still
occasionally run test with GT.M versions as part of developing new tests, make the messages from
both YDB and GTM be excluded by adding back the GTM messages to be excluded.