[YDBTest#470] Remove local copies of cre_coll_sl.csh making tests use com/cre_coll_sl.csh instead
Most of the changes here in the collation and merge tests involve replacing calls to a local version
of cre_coll_sl.csh to the version of this script that lives in the com test directory. Some of the
uses of the local scripts (in both the collation and merge suites) use two different collation
sequences and while the local cre_coll_sl.csh script builds two collation libraries, the generic
version in the com directory does not so we have to call the generic version twice in that case.

This update has a change to the parameters of com/cre_coll_sl.csh. Previously, only part of the name
of the routine to load (only from com/) was specified as the first parameter. Now, that parameter is
the relative load path of the routine wherever it is (no assumptions). So while the previous call
method was something like:

   source $gtm_tst/com/cre_coll_sl.csh polish 1

this is changed to the following:

   source $gtm_tst/com/cre_coll_sl.csh com/col_polish.c 1

Note the collation routine MUST start with the "col_" prefix as that is assumed in the script. Because
of this change, all uses of this script had to be modified. This also allowed all local copies of the
script to be removed.

- endiancvt/cre_coll.csh - was removed as nothing called it any more (was a dead/orphaned script).
- collation/u_inref/polish.c - was removed because the local version was a subset of the com/ version.
- collation/u_inref/polish_rev.c - was removed because the local verson was a subset of the com/ version.

Note that the local cre_coll_sl.csh scripts did not set gtm_local_collate but the new com/cre_coll_sl.csh
scripts do set gtm_local_collate so some tests needed gtm_local_collate explicitly unsetenv'd to pass.

Have run the gde, collation, longstr, merge, spanning_regions, v60003, v62000, and v63013 suites in their
entirety a few times to verify all is working.