[YDBTest#492] [V6.3-006] [GTM-7952] Add v63014/sigsafe test to verify the SIGSAFE external call table options works as described
The SIGSAFE keyword added to external call table entries was added in V63006 but
no tests were written for it for YottaDB. This test corrects that oversight.

Note - the basic part of this test was part of the v63014/gtm9331 test. That part
of the test was severed from gtm9331 and moved to this sigsafe test as the basis
that we added more to. Specifically what was done was:
- gtm9331a stage has been renamed to gtm9331.
- gtm9331b stage has been renamed to v63006/gtm7952

This test has two parts that occur at the same time that verify the SIGALRM function
but only one of those ways actually works on YottaDB. This test defines an external
(C) call that resets the signal handler for SIGALRM. This can cause a test to miss
the SIGALRM wakeup signal resulting in a HANG.

Now both GT.M and YottaDB have code to protect against this action. On return from
a normal external call (not a SIGSAFE call), the existing code extracts the current
signal handler and compares it to what it should be (normal SIGALRM handler). If it
is not that handler, it resets it to that handler and pretends a signal happened and
drives any pending timer that has expired. It also sends a TIMERHANDLER warning to
the system log.

In contrast, when a SIGSAFE enabled external call is made, the code trusts that the
external call won't do any signal handler changes. So on GT.M, when the rules are
not followed and a SIGSAFE declared external call changes the handler, no detection of
the fact is done and if a SIGALRM happened that woke up a timer while the new handler
was in charge results in us not waking up - effectively, the process is hung.

This works differently on YottaDB than it does on GT.M. On YottaDB, a different method
of sleeping is used that does not use SIGALRM so when a SIGALRM handler is stolen, it
does not matter - at least as far as waking up after HANG commands. Other timers that
may exist may not wakeup.

So this test is set up to work regardless of whether the HANGing routine gets hung driving
the non-SIGSAFE routine or not. An extra routine is JOBbed off that will force the main
routine to die if it gets stuck though, again, this is only an issue if the test is run
on GT.M. So the first part of this test just runs this routine through its paces seeing
whether it hangs or not.

The second phase of this this test occurs at the same time. When we do the zsystem to
fork off the processes that actually run the external call and reset the handler, we
actually execute the command with an strace command so the system call trace from the
process is saved.

Later we can go in and grep for all the rt_sigaction(SIGALRM,...) calls that were created
in each of the two runs and note the differences between them - the call not using the
SIGSAFE attribute has 2 more 'rt_sigaction(SIGALRM,...)' calls than the call that had the
SIGSAFE attribute active. The discussion of exactly how all this part works is in a large
block comment in v63014/u_inref/sigsafe.csh.

Note, on the external calls, both calls are the same exact external call but one of them
uses the SIGSAFE attribute in the definition and one does not.