[YDBTest#477] Fix v63014/gtm9285 for ARM processor systems - max shutdown time is different for ARM
The v63014/gtm9285 has a subtest in it that tries to set the soft tries period larger than the
maximum shutdown time which generates an informatory message about it. The issue is that the
maximum shutdown time (which is a system constant) is DIFFERENT for ARM platforms where the
value is doubled. The file YDB/sr_unix/gtmsource.h has the following definition for max shutdown
time:
```C
/* Wait for a max of 2 minutes on a single region database as all the source server shutdown
 * timeouts seen so far have been on a single region database. For multi-region databases, wait
 * for a max of one and a half minute per region. If ever we see timeouts with multi-region
 * databases, this value needs to be bumped as well.
 * On ARM boxes, we have seen cases where the shutdown took a little more than 2 minutes for a single region.
 * So set the maxtimeout of the 1-region case for ARM at 4 minutes.
 */
```
Here's the diff file from such a failure:
```
49,50c49,50
< Soft tries period cannot be more than half of the maximum shutdown wait time of 120 seconds. Reducing soft tries period to 60 seconds
< Soft tries period = 60 seconds, REPLALERT message period = 0 seconds
---
> Soft tries period cannot be more than half of the maximum shutdown wait time of 240 seconds. Reducing soft tries period to 120 seconds
> Soft tries period = 120 seconds, REPLALERT message period = 0 seconds
```
The fix here is to restrict the first message to LINUX_HOST_X86_64 and the second message to
both HOST_LINUX_ARMVXL and HOST_LINUX_AARCH64.

Test now passes on both X86_64 and ARM.