[#495] Rework sudo/plugins test; test --octo, --gui
The sudo/plugins now tests --octo & --gui. However, we found that adding
the two options to the current test framework, which tests every option,
caused the test execution time to balloon from 3 minutes to 41 minutes.

The test was re-written so that it would try 8 random combinations of
options (--plugins-only and --utf8 default are part of that
randomization as well). Now we check to see if any of the installs
failed. The test now takes about 1 minute to complete.