[YDB#780] V7.0-000 [#510] [GTM-9357] Changes to test system caused by removal of ^%DSEWRAP from the test system in favor of ^%PEEKBYNAME()
With this GT.M issue, %DSEWRAP.m is removed from the system so tests that use it must now use ^%PEEKBYNAME(). The test bits changed are:

- com/datinfo.m - Removed ^%DSEWRAP method of obtaining DSE information - simplified routine (tested with refresh_sec/use_backup_and_recover).
- mupip/inref/musetchk.m - Significant changes to use $$^%PEEKBYNAME() instead of ^%DSEWRAP of several fields (see below).
- v60000/inref/searchdump.m - Removed as is no longer used.
- v60000/u_inref/C9B12001861.csh - Changed invocation of searchdump.m to print one field with ^%DSEWRAP to printing one field with ^%PEEKBYNAME().
- v60001/inref/gtm7160.m - Removed as it is no longer used (old DSEWRAP test).
- v60001/outref/gtm7160.txt - removed as it is no longer used (old DSEWRAP test).
- v60001/u_inref/gtm7160.csh - Test removed as it was the old test for DSEWRAP.m which is removed from the test system in favor of PEEKBYNAME.m.
- v60003/gtm7777.m - Changed DSEWRAP usage to fetching fields with ^%PEEKBYNAME().

The changes for mupip/inref/musetchk.m were significant. I tried to make the smallest set of
changes possible by creating the same output that was previous created by the extract from the
output of %DSEWRAP so the extracted before and after files looked the same.

Previously, %DSEWRAP was used to create an array of the possible DSE fields that were displayed.
There was a variable (keep) that listed the field names by which the values generated by %DSEWRAP
were indexed. I kept this array to serve as field titles used later. I added a 'where' array indexed
the same as 'keep' (numeric increasing index) to say what the field name actually was in the
sgmnt_data structure (all fields were from this struct - if any had been from other structs,
the structure name would be part of the fieldname in this array). Since these are the raw values
from the structure, a 3rd array called 'fldfmt' and indexed the same way had either a 0 or a
function name to format the value. If the field was a 0, the raw value was used as-is. Else the
formatter was called. We created the output files just like they were before and the second
part of the test (the checker) was untouched.

BTW, went with 0 instead of null indicating no formatter as it was hard to count ;;;; to figure
out where to insert a field formatter. The 0s made it easier.