[#496] Add v70000/gtm6952 test to validate decimal params can now be specified in hex
This was an encompassing update that affected a LOT of parameters in the utilities (primarily MUPIP
and DSE). Almost any parameter that took a decimal value had the ability added to also accept
hexidecimal values also prefixed by '0x' to indicate the hexadecimal. These changes occurred
primarily in generalized parsing routines, so, for example, a routine that called cli_get_int() to
fetch a 32 bit integer, now calls a routine "cli_is_hex_explicit()" to see if the parm is specified
in hex and if so, parses the value differently.

This test takes a list of those places that call cli_is_hex_explicit() as that is the cornerstone
of this support and selects an example of a command and parms that goes through those places as a
test of all such commands that go through that place. The parsing routines are all well vetted and
this just adds some additional parsing options to get decimal or hex values (which routines already
exist) so we really only need one example from each place. There is one subtest in this test for
each of those places with the exception of the cli_numeric_check() (which also calls
cli_is_hex_explicit()) routine. It turns out that EVERY numeric argument goes through the
cli_numeric_check() routine so we include no additional test for it.

The first subtest needs a bit of additional explanation. One of the review comments was to actually
test a 64 bit value as the value in a `-buffsize=` option when starting the source server. I wanted
to test the hex value on the primary source server and the decimal value on the secondary's passive
source server. This subtest uses the MSR form of replication which, after starting everything with
the `tst_buffsize` envvar set to 0x100001000, I then shut down the secondary and reset the
`tst_buffsize` option to the equivalent decimal value (4294971392) and restart the secondary. Then I
can pull the the command used from the two source logs and get the starting buffsize to show the
64 bit values used. Note that this value is ALSO used for the receive buffer so the systems need to be
able to create 3 of those >4GB buffers. This leaves out the 32 bit ARMV6L systems of course so the
test is restricted there.